+++ lareleve-ejb/src/main/java/com/lr/ejb/AuthenticationEJB.java
@@ -1,7 +1,5 @@
 package com.lr.ejb;
 
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
 import java.util.LinkedList;
 import java.util.List;
 
@@ -55,32 +53,6 @@ public class AuthenticationEJB extends BasicEJB implements IAuthenticationEJBRem
 		} catch (Exception e) {
 			return null;
 		}
-	}
-
-	public String MD5Hashing(String password) {
-
-		try {
-			MessageDigest md = MessageDigest.getInstance("MD5");
-			md.update(password.getBytes());
-
-			byte byteData[] = md.digest();
-
-			//convert the byte to hex format
-			StringBuffer hexString = new StringBuffer();
-			for (int i=0;i<byteData.length;i++) {
-				String hex=Integer.toHexString(0xff & byteData[i]);
-				if(hex.length()==1) hexString.append('0');
-				hexString.append(hex);
-			}
-
-			LOGGER.logDebug(this, "MD5Hashing :", hexString.toString());
-			return hexString.toString();
-
-		} catch (NoSuchAlgorithmException e) {
-			e.printStackTrace();
-		}
-
-		return null;
 	}
 
 }
