+++ lareleve-ejb/src/main/java/com/lr/entity/MD5Hashing.java
@@ -1,0 +1,42 @@
+package com.lr.entity;
+
+import java.security.MessageDigest;
+import java.security.NoSuchAlgorithmException;
+
+public class MD5Hashing {
+
+	private static MD5Hashing instance;
+
+	public static MD5Hashing getInstance(){
+		if(instance==null){
+			instance = new MD5Hashing();
+		}
+		return instance;
+	}
+
+	public String MD5Hashing(String password) {
+
+		try {
+			MessageDigest md = MessageDigest.getInstance("MD5");
+			md.update(password.getBytes());
+
+			byte byteData[] = md.digest();
+
+			//convert the byte to hex format
+			StringBuffer hexString = new StringBuffer();
+			for (int i=0;i<byteData.length;i++) {
+				String hex=Integer.toHexString(0xff & byteData[i]);
+				if(hex.length()==1) hexString.append('0');
+				hexString.append(hex);
+			}
+
+			return hexString.toString();
+
+		} catch (NoSuchAlgorithmException e) {
+			e.printStackTrace();
+		}
+
+		return null;
+	}
+
+}
